==> 02c_bertopic_exploration.ipynb <==
### Load packages

# pip install bertopic
import os
import pandas as pd
import numpy as np
import time

from bertopic import BERTopic
from umap import UMAP
from hdbscan import HDBSCAN
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.decomposition import PCA
from sentence_transformers import SentenceTransformer

### Load data

data = pd.read_csv("data/02b_topic_modeling_prep.csv", sep=",", encoding='utf-8').drop(columns=["Unnamed: 0"])

data.shape

### Set some automatic analytic parameters according to the size of the dataset

samplesize = 1000000
mintopic = int(samplesize/1000)
mintopic

mindfset = int(samplesize/10000)
mindfset

mindfset2 = int(mindfset/10)
mindfset2

### Get a random subsample of the data

honors_docs = data['honors'].astype(str).sample(n=samplesize, random_state=1234).tolist()
details_docs = data['details'].astype(str).sample(n=samplesize, random_state=5678).tolist()

### Initialize some model parameters and convenience functions

def rescale(x, inplace=False):
    """ Rescale an embedding so optimization will not have convergence issues.
    """
    if not inplace:
        x = np.array(x, copy=True)

    x /= np.std(x[:, 0]) * 10000

    return x

sentence_model = SentenceTransformer("all-MiniLM-L6-v2", device="cuda")
hdbscan_model = HDBSCAN(min_cluster_size=mintopic, metric='euclidean', cluster_selection_method='eom', prediction_data=True)
vectorizer_model_1 = CountVectorizer(min_df=mindfset)
vectorizer_model_2 = CountVectorizer(stop_words="english", ngram_range=(1,3), min_df=mindfset2)

### Run BERTopic on the Honors data first

start_time = time.time()
honors_embeddings = sentence_model.encode(honors_docs, show_progress_bar=False)
print("--- %s seconds ---" % (time.time() - start_time))

start_time = time.time()
# Initialize and rescale PCA embeddings
honors_pca_embeddings = rescale(PCA(n_components=5).fit_transform(honors_embeddings))

# Start UMAP from PCA embeddings
honors_umap_model = UMAP(
    n_neighbors=15,
    n_components=5,
    min_dist=0.0,
    metric="cosine",
    init=honors_pca_embeddings,
    low_memory=True
)
print("--- %s seconds ---" % (time.time() - start_time))

start_time = time.time()
# Pass the model to BERTopic:
honors_topic_model = BERTopic(umap_model=honors_umap_model, hdbscan_model=hdbscan_model, vectorizer_model=vectorizer_model_1)
print("--- %s seconds ---" % (time.time() - start_time))

start_time = time.time()
honors_topics, honors_probs = honors_topic_model.fit_transform(honors_docs, honors_embeddings)
print("--- %s seconds ---" % (time.time() - start_time))

start_time = time.time()
honors_topic_model.update_topics(honors_docs, honors_topics, vectorizer_model=vectorizer_model_2)
print("--- %s seconds ---" % (time.time() - start_time))

### Get some of the various output datasets from the run of BERTopic

honors_topic_info = honors_topic_model.get_topic_info()
honors_topic_info

honors_topic_term_matrix = honors_topic_model.c_tf_idf

honors_words = vectorizer_model_2.get_feature_names()

### Prepare and then save the  output

honors_output = pd.DataFrame(honors_topic_term_matrix.toarray())
honors_output.columns = honors_words
honors_output = honors_output.T

honors_output

honors_output.to_csv("02c_honors_topics.csv")

### Run the whole thing again only looking at the Details field data

start_time = time.time()
details_embeddings = sentence_model.encode(details_docs, show_progress_bar=False)
print("--- %s seconds ---" % (time.time() - start_time))

start_time = time.time()
# Initialize and rescale PCA embeddings
details_pca_embeddings = rescale(PCA(n_components=5).fit_transform(details_embeddings))

# Start UMAP from PCA embeddings
details_umap_model = UMAP(
    n_neighbors=15,
    n_components=5,
    min_dist=0.0,
    metric="cosine",
    init=details_pca_embeddings,
    low_memory=True
)
print("--- %s seconds ---" % (time.time() - start_time))

start_time = time.time()
# Pass the model to BERTopic:
details_topic_model = BERTopic(umap_model=details_umap_model, hdbscan_model=hdbscan_model, vectorizer_model=vectorizer_model_1)
print("--- %s seconds ---" % (time.time() - start_time))

start_time = time.time()
details_topics, details_probs = details_topic_model.fit_transform(details_docs, details_embeddings)
print("--- %s seconds ---" % (time.time() - start_time))

start_time = time.time()
details_topic_model.update_topics(details_docs, details_topics, vectorizer_model=vectorizer_model_2)
print("--- %s seconds ---" % (time.time() - start_time))

details_topic_info = details_topic_model.get_topic_info()
details_topic_info

details_topic_term_matrix = details_topic_model.c_tf_idf

details_words = vectorizer_model_2.get_feature_names()

details_output = pd.DataFrame(details_topic_term_matrix.toarray())
details_output.columns = details_words
details_output = details_output.T

details_output

details_output.to_csv("02c_details_topics.csv")
